requirements.txt
groq==0.9.0
supabase==2.9.0
python-docx==1.1.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.6.0
gspread==6.0.0
google-auth==2.27.0
docx2pdf==0.1.8
pypandoc==1.16.2
